Move ob-proxy
