+++ app/index.js
@@ -7,7 +7,7 @@ const bodyParser = require('body-parser');
 const { session } = require('./session');
 const { requireAuthorization } = require('./authorization');
 const { login } = require('./login');
-const { resourceRequestHandler } = require('./ob-proxy.js');
+const { resourceRequestHandler } = require('./request-data/ob-proxy.js');
 const { OBAccountPaymentServiceProviders } = require('./ob-directory');
 const { accountRequestAuthoriseConsent } = require('./account-request-authorise-consent');
 const { authorisationCodeGrantedHandler } = require('./authorisation-code-granted');
